ingestion/requirements-ingestion.txt
google-cloud-bigquery
google-cloud-storage
ijson
chardet
pandas
pyarrow
duckdb
